build(stencil): only export component class from component file
only export component class from component file

COMUI-286
