+++ src/components/beta/gux-tag/gux-tag.types.ts
@@ -1,0 +1,12 @@
+export type GuxTagColor =
+  | 'default'
+  | 'navy'
+  | 'blue'
+  | 'electric-purple'
+  | 'aqua-green'
+  | 'fuscha'
+  | 'dark-purple'
+  | 'bubblegum-pink'
+  | 'olive-green'
+  | 'lilac'
+  | 'yellow-green';
